// frontend/ctrlvalue/src/app/services/api.generated.ts
export interface AccountDto {
    id?: string;
    name?: string | undefined;
    accountType?: string | undefined;
    assetClass?: string | undefined;
    liquidityClass?: string | undefined;
    currency?: string | undefined;
    institution?: string | undefined;
    accountNumber?: string | undefined;
    notes?: string | undefined;
    isActive?: boolean;
    creditLimit?: number | undefined;
    currentBalance?: number;
    startingBalance?: number;
    startingBalanceDate?: Date;
    isOffsetAccount?: boolean;
    openedAt?: Date | undefined;
    closedAt?: Date | undefined;
    externalId?: string | undefined;
    lastSyncedAt?: Date | undefined;
    createdAt?: Date;
    updatedAt?: Date | undefined;
}